[enhancement_getattr] Increment to 0.4.0, adjusted all exposure on highscores as toplevel attributes.
