+++ osrs_highscores/categories.py
@@ -85,7 +85,7 @@ default_boss_ranks = [
     "zulrah",
 ]
 
-default_list = default_optional_ranks + default_boss_ranks
+default_list = default_skills + default_optional_ranks + default_boss_ranks
 
 skill_dict = dict()
 alt_dict = dict()
@@ -135,33 +135,3 @@ class OSRSInfo (object):
         for key, value in target_dict.items():
             info[value["name"]] = key
         return info
-
-
-class OSRSRank(object):
-    def __init__(self, username, rank_type, **kwargs):
-        self.username = username
-        self.type = rank_type
-        self._instantiate(**kwargs)
-
-    def _is_skill(self, **kwargs):
-        try:
-            self.xp = kwargs.get('xp')
-            self.level = kwargs.get('level')
-            self.skill = kwargs.get('skill')
-        except Exception as err:
-            raise err
-
-    def _is_non_skill(self, **kwargs):
-        try:
-            self.score = kwargs.get('score')
-        except Exception as err:
-            raise err
-
-    def _instantiate(self, **kwargs):
-        self.rank = kwargs.get('rank')
-        if self.type == 'skill':
-            self._is_skill(**kwargs)
-        elif self.type == 'nonskill':
-            self._is_non_skill(**kwargs)
-        else:
-            raise ValueError('Target type is not a valid identifier (skill/nonskill')
