+++ osrs_highscores/rankings.py
@@ -1,8 +1,49 @@
 from bs4 import BeautifulSoup
 import requests
 import time
-from .categories import OSRSInfo, OSRSRank
-from .base import OSRSBase
+from .categories import OSRSInfo
+from .base import OSRSBase, with_get_attr
+
+
+@with_get_attr
+class OSRSRank(object):
+    """OSRSRank
+
+    This is a placeholder class for information from a target rank. Returns from Ranking class
+
+    Args:
+        username  str: Username of Account at rank
+        rank_type str: skill/nonskill identifier
+        **kwargs: Additional keywords for attributes to be applied.
+
+    """
+    def __init__(self, username, rank_type, **kwargs):
+        self.username = username
+        self.type = rank_type
+        self._instantiate(**kwargs)
+
+    def _is_skill(self, **kwargs):
+        try:
+            self.xp = kwargs.get('xp')
+            self.level = kwargs.get('level')
+            self.skill = kwargs.get('skill')
+        except Exception as err:
+            raise err
+
+    def _is_non_skill(self, **kwargs):
+        try:
+            self.score = kwargs.get('score')
+        except Exception as err:
+            raise err
+
+    def _instantiate(self, **kwargs):
+        self.rank = kwargs.get('rank')
+        if self.type == 'skill':
+            self._is_skill(**kwargs)
+        elif self.type == 'nonskill':
+            self._is_non_skill(**kwargs)
+        else:
+            raise ValueError('Target type is not a valid identifier (skill/nonskill')
 
 
 class Rankings(OSRSBase):
