+++ requirements.txt
@@ -1,2 +1,3 @@
 requests==2.22.0
 beautifulsoup4==4.8.2
+addict==2.2.1
